Remove CloudBuild script and project folders; Rename test methods; Add NoEditorVR test
