--- Assets.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ff2be840cf2ec1748b4bd144857a31c9
-folderAsset: yes
-timeCreated: 1484170903
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
