--- ProjectSettings.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 6df07c4c145b5f848a37ff7a227f9250
-folderAsset: yes
-timeCreated: 1484170903
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
