--- ProjectSettings/.keep
@@ -1,1 +1,0 @@
-Workaround for CloudBuild automation
