--- ucb_prebuild.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-SCRIPT_DIR=$(cd $(dirname $0) && pwd)
-mv ${SCRIPT_DIR}/* "${SCRIPT_DIR}/Assets/"
