--- ucb_prebuild.sh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8d642107cf4c4c241a9f2d10ebf03521
-timeCreated: 1484170903
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
